requirements.txt
paiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appnope==0.1.3
astor==0.8.1
asttokens==2.4.0
async-timeout==4.0.3
atlassian-python-api==3.41.3
attrs==23.1.0
Authlib==1.2.1
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
camelot-py==0.11.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
ci-info==0.3.0
clarifai==9.8.1
clarifai-grpc==9.8.2
click==7.1.2
coloredlogs==15.0.1
comm==0.1.4
configobj==5.0.8
configparser==6.0.0
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
dotenv-python==0.0.1
EbookLib==0.18
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.0
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.3
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==23.5.26
fonttools==4.43.1
frontend==0.0.3
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.35
googleapis-common-protos==1.60.0
greenlet==3.0.0
grpcio==1.58.0
h11==0.14.0
html2text==2020.1.16
httplib2==0.22.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipykernel==6.25.2
ipython==8.16.1
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.0.310
langdetect==1.0.9
langsmith==0.0.43
layoutparser==0.3.4
llama-hub==0.0.38
llama-index==0.8.48
llama_cpp_python==0.2.2
looseversion==1.3.0
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
olefile==0.46
omegaconf==2.3.0
onnx==1.14.1
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.1.78
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathlib==1.0.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipdeptree==2.13.0
platformdirs==3.11.0
plotly==5.17.0
portalocker==2.8.2
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.3
prov==2.0.0
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pydeck==0.8.0
pydot==1.4.2
Pygments==2.16.1
Pympler==1.0.1
pypandoc==1.12
pyparsing==3.1.1
pypdf==3.15.5
PyPDF2==2.12.1
pypdfium2==4.22.0
PyPika==0.48.9
pyrate-limiter==3.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-rapidjson==1.11
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.6
PyYAML==6.0.1
pyzmq==25.1.1
rapidfuzz==3.4.0
rdflib==7.0.0
referencing==0.30.2
regex==2023.8.8
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.4.2
rpds-py==0.10.2
safetensors==0.4.0
scipy==1.11.3
sec-edgar-downloader==5.0.0
simplejson==3.19.2
six==1.16.0
slack-sdk==3.23.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.2
sympy==1.12
tabula-py==2.8.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
timm==0.9.8
tokenizers==0.14.0
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.11.2
traits==6.3.2
transformers==4.34.1
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.10.25
unstructured-inference==0.7.9
unstructured.pytesseract==0.3.12
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
validators==0.21.0
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.8
weaviate-client==3.23.2
websockets==11.0.3
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.2
zipp==3.16.2
